+++ lib/flop/nimble_schemas.ex
@@ -106,8 +106,10 @@ defmodule Flop.NimbleSchemas do
           type: :keyword_list,
           keys: [
             filter: [
-              type: {:tuple, [:atom, :atom, :keyword_list]},
-              required: true
+              type: {:tuple, [:atom, :atom, :keyword_list]}
+            ],
+            order_by: [
+              type: {:tuple, [:atom, :atom, :keyword_list]}
             ],
             ecto_type: [type: :any],
             bindings: [type: {:list, :atom}],